backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
python-multipart>=0.0.6
httpx>=0.26.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=2.1.0
slowapi>=0.1.9
qdrant-client>=1.7.0
redis>=5.0.1
